Move calcWeight to GHUtility
