+++ core/src/main/java/com/graphhopper/routing/AStar.java
@@ -19,16 +19,12 @@ package com.graphhopper.routing;
 
 import com.graphhopper.coll.GHIntObjectHashMap;
 import com.graphhopper.routing.util.TraversalMode;
-import com.graphhopper.routing.weighting.AbstractWeighting;
 import com.graphhopper.routing.weighting.BeelineWeightApproximator;
 import com.graphhopper.routing.weighting.WeightApproximator;
 import com.graphhopper.routing.weighting.Weighting;
 import com.graphhopper.storage.Graph;
 import com.graphhopper.storage.SPTEntry;
-import com.graphhopper.util.EdgeIterator;
-import com.graphhopper.util.EdgeIteratorState;
-import com.graphhopper.util.Helper;
-import com.graphhopper.util.Parameters;
+import com.graphhopper.util.*;
 
 import java.util.PriorityQueue;
 
@@ -101,7 +97,7 @@ public class AStar extends AbstractRoutingAlgorithm {
                 // todo: for #1776/#1835 move the access check into weighting
                 double tmpWeight = !outEdgeFilter.accept(iter)
                         ? Double.POSITIVE_INFINITY
-                        : (AbstractWeighting.calcWeightWithTurnWeight(weighting, iter, false, currEdge.edge) + currEdge.weightOfVisitedPath);
+                        : (GHUtility.calcWeightWithTurnWeight(weighting, iter, false, currEdge.edge) + currEdge.weightOfVisitedPath);
                 if (Double.isInfinite(tmpWeight)) {
                     continue;
                 }
