+++ core/src/main/java/com/graphhopper/util/details/TimeDetails.java
@@ -17,9 +17,9 @@
  */
 package com.graphhopper.util.details;
 
-import com.graphhopper.routing.weighting.AbstractWeighting;
 import com.graphhopper.routing.weighting.Weighting;
 import com.graphhopper.util.EdgeIteratorState;
+import com.graphhopper.util.GHUtility;
 
 import static com.graphhopper.util.Parameters.Details.TIME;
 
@@ -43,7 +43,7 @@ public class TimeDetails extends AbstractPathDetailsBuilder {
     @Override
     public boolean isEdgeDifferentToLastEdge(EdgeIteratorState edge) {
         if (edge.getEdge() != prevEdgeId) {
-            time = AbstractWeighting.calcMillisWithTurnMillis(weighting, edge, false, prevEdgeId);
+            time = GHUtility.calcMillisWithTurnMillis(weighting, edge, false, prevEdgeId);
             prevEdgeId = edge.getEdge();
             return true;
         }
